# src/frontend/openapi.json
{"openapi":"3.0.0","info":{"title":"Test swagger","description":"Testing the Fastify swagger API","version":"0.1.0"},"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":{}},"paths":{"/":{"get":{"summary":"Welcome endpoint","tags":["Root"],"description":"Returns a welcome message","responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]}}}}}}},"/auth/login":{"post":{"summary":"Login endpoint","tags":["Auth"],"description":"User login","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"userId":{"type":"string"},"password":{"type":"string"}},"required":["userId","password"]}}},"required":true},"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"token":{"type":"string"}},"required":["token"]}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}}}}},"/auth/signup":{"post":{"summary":"Sign up endpoint","tags":["Auth"],"description":"User sign up","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"userId":{"type":"string"},"password":{"type":"string"},"name":{"type":"string"}},"required":["userId","password","name"]}}},"required":true},"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}}}}},"/auth/protected":{"get":{"summary":"Protected endpoint requiring JWT","tags":["Auth"],"description":"Protected endpoint","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}}}}},"/upload/upload-url":{"post":{"summary":"Get pre-signed URL","tags":["Upload"],"description":"Get a pre-signed URL for video uploads","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"}},"required":["title","description"]}}},"required":true},"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"url":{"type":"string"},"videoId":{"type":"string"}},"required":["url","videoId"]}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}}}}},"/upload/transcode-video":{"post":{"summary":"Transcode video","tags":["Upload"],"description":"Transcode a video","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"videoID":{"type":"string"}},"required":["videoID"]}}},"required":true},"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"string"}},"required":["success"]}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}}}}},"/upload/sse/{videoID}":{"get":{"summary":"Get transcoding percentage","tags":["Upload"],"description":"Get transcoding percentage","parameters":[{"schema":{"type":"string"},"in":"path","name":"videoID","required":true}],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Default Response"}}}},"/upload/ws/{videoID}":{"get":{"summary":"Get transcoding percentage","tags":["Upload"],"description":"Get transcoding percentage","parameters":[{"schema":{"type":"string"},"in":"path","name":"videoID","required":true}],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Default Response"}}}},"/user/":{"put":{"summary":"Update user profile endpoint","tags":["User"],"description":"Update user profile","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"},"password":{"type":"string"}}}}}},"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}}}},"get":{"summary":"Get current user","tags":["User"],"description":"Get currrent user","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"userId":{"type":"string"},"name":{"type":"string"}},"required":["userId","name"]}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}}}},"delete":{"summary":"Delete current user","tags":["User"],"description":"Delete current user","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"}},"required":["success"]}}}},"401":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}}}}},"/video/all-videos":{"get":{"tags":["Video"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"userId":{"type":"string"},"title":{"type":"string"},"uploadDate":{"format":"date-time","type":"string"}},"required":["id","userId","title","uploadDate"]}}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}}}}},"/video/{videoId}":{"get":{"tags":["Video"],"parameters":[{"schema":{"type":"string"},"in":"path","name":"videoId","required":true}],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string"},"userId":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"uploadDate":{"format":"date-time","type":"string"},"likes":{"type":"number"},"userHasLiked":{"type":"boolean"},"views":{"type":"number"},"comments":{"type":"array","items":{"type":"object","properties":{"author":{"type":"string"},"text":{"type":"string"}},"required":["author","text"]}}},"required":["id","userId","title","description","uploadDate","likes","userHasLiked","views","comments"]}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}}}}},"/video/search":{"post":{"tags":["Video"],"parameters":[{"schema":{"type":"string"},"in":"query","name":"q","required":true}],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"userId":{"type":"string"},"title":{"type":"string"},"uploadDate":{"format":"date-time","type":"string"}},"required":["id","userId","title","uploadDate"]}}}}},"400":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}},"404":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}}}}},"/video/{videoId}/like":{"get":{"tags":["Video"],"parameters":[{"schema":{"type":"boolean"},"in":"query","name":"isLiking","required":true},{"schema":{"type":"string"},"in":"path","name":"videoId","required":true}],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}}}}},"/video/{videoId}/comment":{"post":{"tags":["Video"],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"comment":{"type":"string"}},"required":["comment"]}}},"required":true},"parameters":[{"schema":{"type":"string"},"in":"path","name":"videoId","required":true}],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{}}}},"500":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}}}}}},"externalDocs":{"url":"https://swagger.io","description":"Find more info here"}}